Rename Python files (PEP8)
