+++ ReadFiles/h52matches.m
@@ -2,7 +2,7 @@ function matchesbis = h52matches(filepath,NbFrames,Version)
 
 % 04/2020 - David Dumont (TO UPDATE)
 %
-% Generates matches array from .h5 file generated by "STM.cpp or STM.py"
+% Generates matches array from .h5 file generated by "STM.cpp or stm.py"
 %% /!\ 2 version of STM.cpp exist and the file are not ordered the same way.
 % For this reason you have to choose the version that correspond to yours
 % in order to see the match.
@@ -20,7 +20,7 @@ if ~exist('Version','var')
     Version='V2';
 end
 
-%% New version 
+%% New version
 % Lecture of tracks
 if Version=='V2'
     matchesbis=[]
@@ -38,5 +38,5 @@ if Version=='V1'
 
 % Lecture of tracks
     matchesbis = h5read(filename,'/tracks');
-end 
+end
 end
